Add files via upload
Read the question and try to do it by yourself.

Code By: RAQEEB RAEES
Create a function that takes an array of integers that represent the amount in dollars that a single stock is worth, and return the maximum profit that could have been made by buying stock on day x and selling stock on day y where y>x.

If given the following array [44,30,24,32,30,40,38,15]then your program should show 16$ profit if person by at 24$ and sells at 40$ which is maximum. And if there is no profit it should return -1.